+++ industry/__init__.py
@@ -1,0 +1,5 @@
+"""Industry research helpers."""
+
+from industry.industry_engine import INDUSTRY_RESEARCH_FIELDS, build_industry_research
+
+__all__ = ["INDUSTRY_RESEARCH_FIELDS", "build_industry_research"]